+++ MigrationProject/include/mrsg.h
@@ -34,6 +34,6 @@ void MRSG_set_task_cost_f ( double (*f)(enum phase_e phase, size_t tid, size_t w
 
 void MRSG_set_dfs_f ( void (*f)(char** dfs_matrix, size_t chunks, size_t workers, unsigned int replicas, int configuration_id) );
 
-void MRSG_set_map_output_f ( size_t (*f)(size_t mid, size_t rid) );
+void MRSG_set_map_output_f ( unsigned long long (*f)(size_t mid, size_t rid) );
 
 #endif /* !MRSG_H */
